Move back tests for Study.optimize to test_study.py
